+++ Packages/com.l47coder.stratum/Templates~/Managers/Asset/AssetManagerData.cs
@@ -1,10 +1,4 @@
-using Stratum;
-
-internal sealed partial class AssetManagerData
+﻿internal sealed partial class AssetManagerData
 {
-    [Field(Readonly = true, Width = 300)]
     public string Key;
-
-    [Field(Readonly = true)]
-    public string AssetAddress;
 }