+++ Packages/com.l47coder.stratum/Templates~/Managers/Task/TaskManagerData.cs
@@ -1,4 +1,4 @@
-internal sealed partial class TaskManagerData
+﻿internal sealed partial class TaskManagerData
 {
     public string Key;
 }